+++ src/Faker/Internal/Types.hs
@@ -74,6 +74,7 @@ data SourceData
   | Fallout
   | FamilyGuy
   | File
+  | FinalSpace
   | Finance
   | Food
   | Football